# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Council Asset & Compliance Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed

  # ── TENANTS (Super Admin) ─────────────────────────────────────────────────
...
  /compliance-types:
    get:
      operationId: listComplianceTypes
      tags: [compliance-types]
      summary: List compliance types for current tenant
      responses:
        "200":
          description: Compliance types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceType"
    post:
      operationId: createComplianceType
      tags: [compliance-types]
      summary: Create a compliance type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceTypeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceType"
  /compliance-types/{typeId}:
    patch:
      operationId: updateComplianceType
      tags: [compliance-types]
      summary: Update a compliance type
      parameters:
        - name: typeId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceTypeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceType"
  /compliance-types/{typeId}/custom-fields/{fieldKey}:
    patch:
      operationId: renameCustomField
      tags: [compliance-types]
      summary: Rename a custom field label on a compliance type
      parameters:
        - name: typeId
          in: path
          required: true
          schema: { type: string }
        - name: fieldKey
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFieldLabelUpdate"
      responses:
        "200":
          description: Updated compliance type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceType"
    delete:
      operationId: deleteCustomField
      tags: [compliance-types]
      summary: Remove a custom field definition from a compliance type
      parameters:
        - name: typeId
          in: path
          required: true
          schema: { type: string }
        - name: fieldKey
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated compliance type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceType"

  # ── ASSET COMPLIANCE ITEMS ────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    UserProfile:
      type: object
      required: [id, username, email, role, tenantId]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, tenant_admin, compliance_manager, team_member, auditor]
        tenantId:
          type: ["string", "null"]
        tenantName:
          type: ["string", "null"]

    # Tenants
...
    CustomFieldDefinition:
      type: object
      required: [key, label]
      properties:
        key:
          type: string
        label:
          type: string
    CustomFieldLabelUpdate:
      type: object
      required: [label]
      properties:
        label:
          type: string
    ComplianceType:
      type: object
      required: [id, name, code, isActive]
      properties:
        id:
          type: string
        tenantId:
          type: ["string", "null"]
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        frequencyMonths:
          type: ["integer", "null"]
        dueSoonDays:
          type: integer
        isActive:
          type: boolean
        isSystem:
          type: boolean
        color:
          type: ["string", "null"]
        applicableAssetTypes:
          type: array
          nullable: true
          items:
            type: string
        customFieldDefinitions:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/CustomFieldDefinition"
    ComplianceTypeInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        frequencyMonths:
          type: integer
        dueSoonDays:
          type: integer
        color:
          type: string
    ComplianceTypeUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        frequencyMonths:
          type: integer
        dueSoonDays:
          type: integer
        isActive:
          type: boolean
        color:
          type: string
        applicableAssetTypes:
          type: array
          nullable: true
          items:
            type: string
...